Revert "Feature/layout"
