+++ client/src/App.js
@@ -28,6 +28,7 @@ const authLink = setContext((_, { headers }) => {
   };
 });
 
+
 const client = new ApolloClient({
   link: authLink.concat(httpLink),
   cache: new InMemoryCache(),
@@ -48,4 +49,3 @@ function App() {
 }
 
 export default App;
-
