--- client/src/components/WorkoutTable.js
@@ -1,1 +1,0 @@
-
